Join viewsets in a single file
